refactor: 새로운 매핑 방법을 적용하여 일부 컨트롤러 삭제 및 UserController 생성
