--- slipp/src/main/java/slipp/controller/CreateUserController.java
@@ -1,29 +1,0 @@
-package slipp.controller;
-
-import javax.servlet.http.HttpServletRequest;
-import javax.servlet.http.HttpServletResponse;
-
-import nextstep.mvc.tobe.JspView;
-import nextstep.mvc.tobe.ModelAndView;
-import nextstep.web.annotation.Controller;
-import nextstep.web.annotation.RequestMapping;
-import nextstep.web.annotation.RequestMethod;
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
-import slipp.domain.User;
-import slipp.support.db.DataBase;
-
-@Controller
-public class CreateUserController {
-    private static final Logger log = LoggerFactory.getLogger(CreateUserController.class);
-
-    @RequestMapping(value = "/users/create", method = RequestMethod.POST)
-    public ModelAndView createUser(HttpServletRequest req, HttpServletResponse resp) throws Exception {
-        User user = new User(req.getParameter("userId"), req.getParameter("password"), req.getParameter("name"),
-                req.getParameter("email"));
-        log.debug("User : {}", user);
-
-        DataBase.addUser(user);
-        return new ModelAndView(new JspView("redirect:/"));
-    }
-}
