--- slipp/src/main/java/slipp/controller/ProfileController.java
@@ -1,34 +1,0 @@
-package slipp.controller;
-
-import javax.servlet.http.HttpServletRequest;
-import javax.servlet.http.HttpServletResponse;
-
-import nextstep.mvc.tobe.JspView;
-import nextstep.mvc.tobe.ModelAndView;
-import nextstep.web.annotation.Controller;
-import nextstep.web.annotation.RequestMapping;
-import nextstep.web.annotation.RequestMethod;
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
-import slipp.domain.User;
-import slipp.support.db.DataBase;
-
-@Controller
-public class ProfileController {
-    private static final Logger log = LoggerFactory.getLogger(ProfileController.class);
-
-    @RequestMapping(value = "/users/profile", method = RequestMethod.GET)
-    public ModelAndView getProfile(HttpServletRequest req, HttpServletResponse resp) throws Exception {
-
-        String userId = req.getParameter("userId");
-        User user = DataBase.findUserById(userId);
-        if (user == null) {
-            throw new NullPointerException("사용자를 찾을 수 없습니다.");
-        }
-
-        log.debug("User: {}", userId);
-
-        return new ModelAndView(new JspView("/user/profile.jsp"))
-                .addObject("user", user);
-    }
-}
