--- slipp/src/main/java/slipp/controller/UserFormController.java
@@ -1,18 +1,0 @@
-package slipp.controller;
-
-import javax.servlet.http.HttpServletRequest;
-import javax.servlet.http.HttpServletResponse;
-
-import nextstep.mvc.tobe.JspView;
-import nextstep.mvc.tobe.ModelAndView;
-import nextstep.web.annotation.Controller;
-import nextstep.web.annotation.RequestMapping;
-import nextstep.web.annotation.RequestMethod;
-
-@Controller
-public class UserFormController {
-    @RequestMapping(value = "/users/form", method = RequestMethod.GET)
-    public ModelAndView getJoinForm(HttpServletRequest req, HttpServletResponse resp) {
-        return new ModelAndView(new JspView("/user/form.jsp"));
-    }
-}
