--- slipp/src/test/java/slipp/controller/UpdateUserControllerTest.java
@@ -1,55 +1,0 @@
-package slipp.controller;
-
-import java.time.Duration;
-
-import org.junit.jupiter.api.BeforeEach;
-import org.junit.jupiter.api.DisplayName;
-import org.junit.jupiter.api.Test;
-
-import org.springframework.http.ResponseCookie;
-import org.springframework.test.web.reactive.server.WebTestClient;
-import org.springframework.web.reactive.function.BodyInserters;
-
-class UpdateUserControllerTest {
-    private static final String BASE_URL = "http://localhost";
-
-    private int port;
-    private WebTestClient webTestClient;
-
-    @BeforeEach
-    void setUp() {
-        this.port = 8080;
-        this.webTestClient = WebTestClient
-                .bindToServer()
-                .baseUrl(BASE_URL + ":" + port)
-                .build();
-    }
-
-    @Test
-    @DisplayName("로그인 시 회원 정보 수정")
-    void update_user() {
-         webTestClient.post()
-                .uri("/users/update")
-                .body(BodyInserters.fromFormData("userId", "admin")
-                        .with("password", "password").with("name", "자바지기")
-                        .with("email", "admin@slipp.net"))
-                .cookie("JSESSIONID", getCookie().getValue())
-                .exchange()
-                .expectStatus().isFound();
-    }
-
-    private ResponseCookie getCookie() {
-        return webTestClient.mutate()
-                .responseTimeout(Duration.ofMillis(70000))
-                .build()
-                .post()
-                .uri("/users/login")
-                .body(BodyInserters.fromFormData("userId", "admin")
-                        .with("password", "password"))
-                .exchange()
-                .expectStatus().isFound()
-                .expectHeader().valueMatches("location", "/")
-                .returnResult(String.class)
-                .getResponseCookies().get("JSESSIONID").get(0);
-    }
-}
